Skip low priority lonely plans if possible
Lonely plans which split input depth are always going to be worse than
plans that don't (for regular convolution). Currently we still generate
these plans and only when we come to estimate them do we realize that
they are worse. This means we spend a lot of time encoding weights for
plans which we then discard.

This patch avoids generating these plans in the first place, and instead
only generates them when none of the other (more efficient) plans
work.

This leads to improved network compilation speed.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>