Move scheduling to support library
Previously, the command stream contained dependency information for all
the agents and the firmware was responsible for using this to schedule
commands on the hardware. This has been sped up by moving this logic to
the support library, such that the command stream now contains lists of
commands rather than dependency information.

Add Scheduler.hpp/cpp file to the support library which roughly does the
same as the scheduler in the firmware used to. It takes as input a list
of agents and their dependencies, and outputs the four command lists.
These are then stored into the command stream.

Remove dependency information (ratios etc.) from command stream. This is
now resolved in the support library and so the firmware does not need
this information. Dependency information between command queues(e.g. an
MCE command that needs to wait for a DmaRd command to finish) are still
contained in the command queues themselves as WaitForAgent commands.

Add cascading "Commands" to command stream, stored as four arrays
(DmaRd, DmaWr, Mce, Ple) that come after the array of agents. This name
was chosen to be consistent with the existing name in the firmware, and
it fits with the name "Command Stream", as these new Commands are now
the main part of the Command Stream.

Each Command is simply an agent ID and stripe ID, as well as a type enum
that specifies what kind of command it is (e.g. a command to program
the MCE registers, or a command to wait for another agent).

Add tests for scheduling logic in the support library.

Update binary-to-xml conversion to support the updated command stream
format.

Improve binary-to-xml unit tests to dump files when there is a
difference, for easier comparison.

The firmware no longer has limitations for the maximum number of MCE or
IFM stripes per PLE stripe, so this has been removed from the
capabilities and support library.

Now that dependency information (ratios etc.) is not stored in the
command stream and is only used as an intermediate step in the support
library, the limitations on the max number of dependencies have been
lifted (they are now std::vectors rather than std::arrays).

Add tests for new scheduler class, based on unit tests from the
firmware's scheduler.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>