Avoid unused identity MCE/PLE plans
Previously, the McePart generated two sets of plans, which placed the
output buffer in both PleInputSram and regular Sram. This was necessary
because we didn't know if the consuming part(s) would be able to take
their input directly from PleInputSram or not. To avoid these extra
plans leading to too many combinations being considered, there was some
logic in the combiner that disallowed combining a plan with an identity
PLE at the end with a plan that had an identity MCE at the start. This
means that some of the plans generated would have been immediately
rejected and so the computational overhead was not that much.

Now that we are looking at cascading over branches, the consuming part
may not be the one that is considered next, and so the extra plans can
lead to an exponentially increasing number of combinations being
considered by the combiner before they get filtered out.

This patch gives parts the information that they need to generate only
plans with outputs in PleInputSram if they can actually be used. This
is done as a step after creating the GraphOfParts, because we need to
know the connections.

This also avoids the need for the logic in the combiner to check if
plans should be rejected during continue and end section plan
generation, which simplifies the code.

Remove "McePart GetPlans InputSramBuffer" test case as it appears to be
a duplicate of "McePart GetPlans structure".

Signed-off-by: Rob Hughes <robert.hughes@arm.com>