Fix unnecessary object copies in support library
There are several instances in the support library where unnecessary
copies are being made of objects and in a few of these instances the
objects are quite large.

To avoid these copies, the instances have been changed to either use
move semantics or const references according to what is applicable for
the situation.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>