Remove find before insert in Visualisation
Currently, to avoid inserting a new value in the map of agent ID ranges
for a pass, the whole map is first searched for the value's key and only
if the key was not found, the value is added with the map's subscript
operator.

Searching the map in this way is unnecessary because the map's insert
function will not add a new value if the specified key already exists so
the logic above has been replaced with calling the insert function
instead.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>