Fix issue with networking splitting and caching in the backend
A network can sometimes be split because of unsupported operations and
one of the splits can't be added to the network. E.g. a constant layer
with batches. This layer will fail to be added to the cache but the
subgraph number will still be incremented in Arm NN.

This causes the subgraph indexes to become out of sync between Arm NN
and what's in the cache.

Solved this by changing the cache to use a map with the subgraph index
as the key.

The cache's format has now changed to store this index. This is not
backwards compatible.

Signed-off-by: Joshua Slater <joshua.slater@arm.com>