Set boundary requirements before plan generation
Previously, parts would generate multiple plans with a variety of
numbers of output stripes in the tile. This is because the following
part may require boundary data, but we didn't know whether boundary
data was needed or not so we generated several plans, and relied on the
following part to reject any plans that didn't match its expectations.
This worked fine because any excess plans were immediately rejected and
so the computational overhead was not that much.

Now that we are looking at cascading over branches, the consuming part
may not be the one that is considered next, and so the extra plans can
lead to an exponentially increasing number of combinations being
considered by the combiner before they get filtered out.

This patch gives parts the information that they need to generate only
plans with the required number of stripes for boundary data. This is
done as a step after creating the GraphOfParts, because we need to know
the connections.

This also avoids the need to check which plans should be rejected during
continue and end section plan generation, which simplifies the code.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>