Remove schedule dependencies
Get rid of scheduling dependencies and instead use the read and write
dependencies to derive the correct scheduling order. This is done by
walking backwards from OfmS stripes using read and write dependencies
to determine what needs to be completed first.

This is less confusing and simplifies the dependency generation. It also
fixes some issues with the old algorithm that show up when attempting
to cascade over branches (it schedules PLE stripes out of order
w.r.t. MCE stripes).

Add checks for stripe ids being in range. This is useful for finding
dependency bugs (e.g. when the dependency ratios are invalid).

Some tweaks were needed to the dependency generation code in
CascadingCommandStreamGenerator.cpp so that the new scheduling
algorithm produces the same results as the old one.

Specifically, the "fence op" concept was modified so that there is now
one common fence op rather than one per agent type. This is simpler and
fixes issues where the PleL was running too soon (on the previous
cascade) and then the PLE code was getting overwritten before being
used.

The order that some dependencies are added has also been tweaked, as the
order of dependencies is now important due to the depth-first traversal
of the dependency graph (whereas previously it did not matter).

Signed-off-by: Rob Hughes <robert.hughes@arm.com>
Signed-off-by: Joshua Slater <joshua.slater@arm.com>