Make firmware and PLE source code public
Scylla register definitions, Firmware and PLE components are made
public.These components were previously not part of the released
driver stack, but are now being released as open source.

The control unit and ple folders are now moved to the public folder
"firmware/". The build scripts, symbolic links and README files are
updated to reflect this.

The following minor bug fixes are also included in this patch.
1. Fix for PLE clusted build issue.
2. Removal of MRI debug monitor support to debug firmware.

Signed-off-by: Joshua Slater <joshua.slater@arm.com>
Signed-off-by: Rob Hughes <robert.hughes@arm.com>
Signed-off-by: Zingo Andersen <zingo.andersen@arm.com>