Detect and raise errors for scheduling deadlocks
If there are bugs in the dependency generation code then the scheduler
can add WaitForAgent commands to the command queues which will never be
satisfied and cause hang in the firmware when running the inference.

This patch adds some checks to detect this in the support library to
make these bugs easier to find and diagnose.

It also skips adding WaitForAgents on stripes which are earlier in the
same queue, and therefore always satisfied.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>