Replace WaitForAgent with WaitForCounter
Rather than waiting for agent stripes to be completed, we instead store
four counters to indicate the progress of the different command queues
and wait for particular values of these counters. This is a more
flexible approach as it allows waiting for things which aren't an agent
stripe, and will be used in a future patch to simplify the MCEIF
handling. It also simplifies the firmware.

Agent struct simplified as no longer need numStripesTotal. stripeId
fields removed from most Command types as no longer needed.
m_IsLastChunk removed from DmaCommand as no longer needed.

Scheduler keeps track of what counter values correspond to what agent
stripes, and puts these in WaitForCounter commands.

Fix issue with logging code not optimised out (use LoggingString)

Rename IsEmpty -> IsDone as it's no longer just queues

Remove agent progress from Controller as no longer needed. Instead we
store the new counters.

Remove callback queue as the actions taken are simply incrementing the
appropriate counter, so the callback queue is overkill for this.

Rename HandleCallbacks to UpdateProgress as there are no more
callbacks.

Remove HwController, as not much logic is now shared, and the
PleController is different (no HW queue)

HwAbstraction::GetNumCmdsInPleQueue replaced with IsPleBusy, as PLE has
no queue

UpdateTsuEvents returns flag to indicate if PLE is finished

Profiling information is unfortunately now more limited as there is less
knowledge of individual stripes in the firmware. This can be remedied
by some post-processing of the profiling data.

Remove RingArray.hpp as not used. Remove TriviallyDestructibleQueue as
no longer used. Remove WindowArray as no longer used.

Remove a lot of SchedulerTests as these are difficult to maintain and
offer limited value now that we have ggf testing.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>