Merge command stream extra data into the commands
Rather than having separate arrays of extra data which are referred to
from Commands, the Commands now contain all their extra data. This
makes it possible to have Commands which don't have an agentId and
stripeId, which will be needed for future changes.

Command is now a base class which contains just a type enum, and there
are multiple sub-types of Command which contain additional fields,
depending on the type. This means the command lists in the command
stream have commands of varying length, so they are no longer regularly
spaced arrays.

The Scheduler in the support library now fills in all the fields, rather
than leaving this until a later step.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>