Preload weights from later layers
Before this patch, the weights, IFMs and PLE kernels for a layer were
never loaded earlier than the start of the current cascade
(using the "fence op" to mark this). This is correct, but not optimal,
because in some cases we can "preload" data for later layers before the
previous cascade has finished, as long as the data being loaded does
not overwrite any data still in use by the previous cascade. This patch
enables this optimisation.

Removing the fence op dependencies means we need to tweak some other
dependencies to get the PleL agents to be in the right order.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>