Automatically remove more redundant WaitForCounter
Adding a WaitForCounter will often mean implicitly waiting for other
counter values too. A trivial example of this would be that waiting for
DmaRd=2 also means waiting for DmaRd=1, but there are more complicated
examples like waiting for MceStripe=2 where the MCE queue waits for
DmaRd=1 before kicking off stripe number 2, means that you are
implicitly waiting for DmaRd=1 as well. This patch adds code to store
these dependencies/implications, and allows us to omit some
WaitForCounters which we can guarantee will always be met.

This results in smaller command stream which will be faster for the
firmware to process, and should have no effect on the correctness of
the command stream.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>