Alternate SRAM allocation preferences
To get more benefit from preloading of weights from later layers, we
want to minimise the overlap of SRAM buffers with earlier buffers, even
when those buffers are no longer being used. This will allow the
command stream generation to detect that the later buffers can be
loaded early, leading to faster inferences.

This patch implements a very simple strategy to achieve this, which is
to alternately allocate new buffers at the start and end of the
available SRAM space.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>