Use thread pool for weight encoding parallelism
The old method of parallelising weight encoding using std::async was
inefficient when running on some C++ implementations (notably gcc) as
it spawns an entire new thread for each task. This can result in worse
performance as there is an overhead in spawning each thread and too
much parallelism can result in thrashing.

Even though we currently only launch a maximum of 32 tasks at the same
time, this might already be too much for some systems, and future
patches will make other work in the support library more parallel,
which would worsen this problem.

Instead of using std::async, we now have our own ThreadPool class which
manages a set of background threads which execute tasks from a queue.
This can be used to run work in the background without blocking the
current thread. The weight encoder uses this. The weight encoder has
been refactored to standalone functions rather than a class to make
this easier to implement. The weight encoder now separates encoding
into two phases (stage 1 and stage 2) which are exposed through its
public API, allowing callers to run multiple stage 1 encodings in
parallel with each other.

The serialized weight encoder cache has been removed as weight encoding
will soon be made much faster and so the benefit of the serialized
cache is much less and the maintenance is no longer worth it.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>