Improve accuracy of bytesNeededToStartProcessing
Previously we were using m_SlotSizeInBytes, but the amount of data
actually transferred can be less than this in cases of partial stripes
(in any of the three dimensions), because the slot size will be for the
full stripe shape and so this would overestimate.

The new updated calculation takes this into account, which leads to the
metric being more accurate when double-buffering is possible, which
leads to plans using double-buffering being chosen more often, leading
to better inference performance.

This is particularly noticeable for the first layer of many networks,
where the number of channels can be 3, which we were previously
treating as 16 and thus overestimating.

Also remove unused legacy functions.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>