Account for block size in PLE estimation
When estimating the performance of the PLE, round up the tensor shape to
the block config, because the PLE always processes an entire block,
even if it is only partial.

This improves the accuracy of the PLE performance estimation, which leads
to better decisions about which block config to choose, leading to
improved inference performance.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>