Pass Ops and Buffers into metric estimation code
The code which estimates a cycle count (metric) for a pass used to use
the same inputs as the SPA estimation (i.e. the PassPerformanceData
struct). This has limited information (e.g. no knowledge of MCE
stripes) which means that the accuracy of the estimated metric is
limited. This patch aims to lift this limitation by passing in
additional information in the form of Ops and Buffers.

The existing PassPerformanceData struct can't be augmented because it is
part of the public API and needs to keep compatibility with old SPA
releases.

Instead, this data has been kept around in a "legacy" variable and the
new Buffers and Ops have been added to a "PassDesc" struct. As this new
struct is internal to the support library, we can add any information
we want to it.

Note that the new information isn't yet used, but following patches
will make use of it.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>