Fix decision for packed boundary data
We were choosing to use packed boundary data in a case where we
shouldn't have been - depthwise with height and depth split, as the
boundary data in this case can be regular stripes rather than packed.

To help reproduce the bug, add the ability to specify the number of
IFM/OFM/weight stripes in the stripe config file.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>