Adjust overhead calculations in performance metric
This is more accurate to the newer firmware, and gives better inference
performance as the support library picks better plans.

This makes use of the new Ops and Buffers passed in to the estimation
code.

Add debugging information to the dot files to show how the metric is
calculated, which is useful for debugging and analysis.

Remove some tests which would have needed updating, but have been become
mostly defunct.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>