Fix clamping of IFM tiles and boundary stripes
The optimisation which shrinks the IFM tile to the size of the tensor
was being overly optimistic and shrinking it too much, in cases where
the IFM is split in multiple dimensions. The code has been re-rewritten
to account for this.

A related issue with boundary stripes has also been fixed, where we were
being too conservative with the amount of boundary data we required.
This should allow less SRAM to be used, and the firmware to be less
restricted by dependencies, and this should improve inference
performance.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>