Make dependency tile size more flexible
Replace the flag writesToTile with an integer writesToTileSize, which
allows setting the tile size to use for dependency calculations. This
means the dependency generation code can be more flexible in the
dependencies it generates.

For now, we are always setting this to a value that results in the same
behaviour as before, but this allows us to change this in future if
necessary. It also simplifies the Scheduler code as it can simply read
this value, rather than have to determine an appropriate tile size to
use.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>