Improve hardware error handling and reporting
When the firmware or hardware crashes, we currently dump out the GP
registers which contain some useful information put there by the
firmware. This patch expands on that concept to include more
information and display it in a friendlier way.

Configure the NPU the send an interrupt to the NCU MCU when the NPU has
an error. The interrupt handler for this dumps information to the GP
registers and then sends an interrupt to the host system. This is the
same process as happens for MCU faults (e.g. invalid memory access).
This allows us to dump the NPU error cause information from the
firmware, which is not accessible by the kernel.

Expand the data dumped into the GP registers by the firmware to include
the exception number that triggered the fault (e.g. a MemManage fault,
HardFault, NPU fault etc.), the TOP_ERR_CAUSE/ADDRESS register and the
CE_ERR_CAUSE/ADDRESS register. Improve the display of this data in the
kernel module prints.

Remove the TOP_ERR_CAUSE checking code from the firmware, and this is no
longer necessary (the firmware will already have gone into the
interrupt handler).

Disable error interrupts being sent to the host when the hardware runs
into errors. Note these are different to the error interrupt sent by
the firmware, which is still enabled. The firmware handles hardware
errors itself and then informs the host via the error interrupt, so the
kernel will still get informed but after the firmware has dumped some
useful debugging information. Technically we could leave these
interrupts enabled and then the host would get informed as well as the
firmware being informed, and there are some issues with this:

 1 - the kernel might see the HW error before the firmware has had a
 chance to dump more detailed error information.

 2 - this seems to result in the host being spammed with thousands of
 interrupts resulting in a hang. It's not clear why this is happening.

Note that "unrecoverable" errors (see HW error categorisation
definitions) are left enabled as these cannot be handled by the
firmware.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>