Improvements to profiling
Halve the size of the profiling entry struct (16 bytes down to 8 bytes).
This makes profiling take up less memory and it should also add less
overhead. This has been achieved by shrinking the timestamp to 32-bits
(and correcting for the resulting potential overflow), removing a lot
of metadata that is no longer necessary(and reconstructing this in the
offline conversion) and by re-using ID values so that we can shrink
this field to 5 bits (down from 16).

Reuse profiling entry IDs once the end event has been sent. This means
we can reduce the size of the ID field and no longer have to be
concerned with ID numbers overflowing (previously there was a soft
limit of 64k entries, after which there were problems).

Remove unused entry types (leftover from legacy and strategy X firmware
code) and consolidate the remaining ones.

Expose firmware label metadata in the public API

Improve profiling JSON dump to include counter names.

Simplify firmware profiling offset exposed from the kernel module, by
making the firmware reset the PMU to zero when configuring profiling.
This avoids the need for the firmware to send back a timestamp.

Move timestamp correction logic from KmodNetwork to common code, so it
can be shared with the model backend.

Because the event IDs from the firmware are now re-used, the driver
library now re-assigns IDs to entries so that each entry exposed from
the driver library API still has a unique ID. This also means we no
longer need to offset the ID for events originating from the driver
library.

Account for overflow of the 32-bit timestamp counter in the driver
library. Add new TimestampFull entry type to allow the firmware to send
a full 64-bit timestamp after a period of inactivity. This is necessary
now that the regular timestamps are only 32-bit to ensure the timestamp
correction doesn't "skip" time.

Simplify most profiling events by removing their extra metadata. This is
no longer needed as can be reconstructed later by the offline
conversion step. This saves some bits and simplifies the firmware and
driver library.

Add ETHOSN_NUMERIC_CAST macro to help with bitfield masking to avoid
compiler warnings.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>