Fix dual core carveout plus FW boot improvements
Carveout with dual core was previously broken because both cores were
using the same stacks and conflicted with each other. This patch makes
it so that each core has a separate stack.

The code parts of the firmware (privileged, unprivileged and PLE code)
are still shared between cores though as this allows us to avoid
compiling with position-independent code. We avoid position-independent
code because it makes the compiled firmware code more complicated and
slower, and the compiler generates code which offsets global variables
with function addresses and this causes problems when resetting the NPU
as the same addresses get offset again. Note that for _SMMU_ dual core,
each core has a completely separate address space and so have their own
copy of the code.

This means that the g_Unprivileged/PrivilegedStack variables and the
vector table's first entry will always point to the first core, which
will be wrong when running the second core, but we fix this by
overwriting it in the kernel module as part of booting the firmware.

The MPU setup is problematic for dual core with carveout, because the
vector table and stacks for the second core will be after the
privileged stack for the first core. This means that they fall into
region 1 which is Read-Write and thus not ideal. There isn't an easy
fix for this so we accept the reduced security as carveout is already
very limited in its security.

Fix a mistake in the MPU diagram.

Call main() rather than __main() (which is the libcxx startup sequence)
because this performs a bunch of initialisation which we don't need or
want because it tries to zero out some memory which is read-only and
will already be zeroed in the firmware binary. This means we can enable
the MPU earlier and mark the firmware code as read-only in the SMMU
mappings.

Improve the reporting of successful firmware boot by moving it as late
as possible in the boot sequence, so that any problems with starting
the unprivileged task are caught. Also replace the setting of a GP to
zero, with setting it to a special value, to make it more robust.

Use ETHOSN_REGION_MASK rather than repeating the calculation.

When the firmware fails to boot or errors, print the value of SYSCTLR0
as this can contain useful information (e.g. if an MCU lockup
occurred).

When loading the firmware for the second core in carveout, skip loading
the code parts of the binary and just load the stacks and vector
table.

When the NPU is rebooted for the second (and later) times, the firmware
is already loaded into memory and so doesn't need to be re-loaded. This
speeds up resetting due to power management, and simplifies the code.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>