Max pooling stride 1 for reasonable kernel sizes
Add a PLE kernel which can do 1D max pooling (in either X or Y) with
stride 1 for a runtime-configurable pooling size and padding(up to some
limits). The support library uses this to run 2D max pooling by
decomposing it into 2 x 1D max poolings (in each of X and Y).

The MaxPool1D kernel supports a pooling size up to 9 for VALID padding,
and 17 for SAME padding. Other padding amounts (neither SAME nor VALID)
are supported by the kernel, but not currently supported by the support
library (to keep testing etc. simple).

The MaxPool1D kernel does not use the common scratch register layout as
we need different parameters.

Update the plan generation logic in StandalonePlePart.  To keep things
simple, for now we don't support any splitting in the direction of the
pooling. This avoids tricky cases of handling extra IFM stripes with
valid padding and managing leftover groups in the PLE kernel. It does
limit the maximum tensor size we can support in that dimension(and this
is part of the supported checks), but the limit is pretty high.

Remove unused inputDepthOnly flag in StripeConfig.

Signed-off-by: Rob Hughes <robert.hughes@arm.com>