Add the ability to use llvm-embedded toolchain
Have the ability to use the open source llvm-embedded toolchain to build
the ple kernels and control unit firmware.

To use it please specify the following SCons parameters or place them in
your options.py
 * use_llvm_embedded=1
 * llvm_embedded_toolchain_path=<path_to_llvm_binaries>

Please note the performance of this toolchain is much worse than the
default armclang toolchain.

Signed-off-by: Joshua Slater <joshua.slater@arm.com>