kernel_module: Move call to ethosn_set_power_ctrl
When the ethosn driver is unloaded the function ethosn_set_power_ctrl
is called from ethosn_device_release. The problem is that
ethosn_device_release is called after ethosn_pdev_remove. When
ethosn_pdev_remove is done the kernel considers the device to not be in
use anymore meaning that it is safe to cut the power to the device. As
a consequnce ethosn_set_power_ctrl will try to write to a register of a
device that is not guaranteed to have power which may lead to a crash.

Change so that ethosn_set_power_ctrl is called at the end of
ethosn_pdev_remove instead of ethosn_device_release.

Signed-off-by: Martin Danielsson <martindn@axis.com>