Move debug formatting to where it is used
CalculateMetric() created string formatting of debug data. This function
is called in the inner loop of the compilation of networks so this is
very costly. The debug data is used elsewhere (and in many cases not at
all). Change so we only store the debug-data in the inner loop and then
do the string formatting where it is actually needed.

This saving trims off 20% compilation time on mobilenet and 10% on
efficientnet.

Signed-off-by: Björn Ardö <bjorn.ardo@axis.com>